requirements.txt
numpy
pandas
h5py
numba
psutil
memory_profiler
pyarrow
tqdm
scipy==1.12.0
python-Levenshtein
joblib